Fix issue with Namespace in same session
Fixed an issue with the namespace when being run multiple times in the same session. For example as part of the release build we deploy multiple different report projects one after another. The first report was deploying fine but reports are that were getting an error:

##[error]Microsoft.PowerShell.Commands.WriteErrorException: Cannot convert argument "DataSources", with value: "SSRS.DataSource", for "SetItemDataSources" to type "SSRS.DataSource[]": "Cannot convert the "SSRS.DataSource" value of type "SSRS.DataSource" to type "SSRS.DataSource[]"."

So I have found this article which seems to have addresses the issue.

http://www.sqlmusings.com/2012/02/04/resolving-ssrs-and-powershell-new-webserviceproxy-namespace-issue/

Thanks for a good tool to deploy our builds.